Renamed _metadata to _units
