+++ esmvalcore/preprocessor/__init__.py
@@ -17,13 +17,13 @@ from ._io import (_get_debug_filename, cleanup, concatenate, load, save,
 from ._mask import (mask_above_threshold, mask_below_threshold,
                     mask_fillvalues, mask_inside_range, mask_landsea,
                     mask_landseaice, mask_outside_range)
-from ._metadata import convert_units
 from ._multimodel import multi_model_statistics
 from ._reformat import (cmor_check_data, cmor_check_metadata, fix_data,
                         fix_file, fix_metadata)
 from ._regrid import extract_levels, regrid
 from ._time import (annual_mean, extract_month, extract_season, extract_time,
                     regrid_time, seasonal_mean, time_average)
+from ._units import convert_units
 from ._volume import (average_volume, depth_integration, extract_trajectory,
                       extract_transect, extract_volume)
